+++ identityhub-service/src/main/java/br/dev/andrestamatto/identityhub/clientapplication/application/ConfidentialClientSecretRotator.java
@@ -1,6 +1,6 @@
 package br.dev.andrestamatto.identityhub.clientapplication.application;
 
-public interface BffClientSecretRotator {
+public interface ConfidentialClientSecretRotator {
 
     ConfidentialClientSecret rotate(ApplicationClientSnapshot client);
 }