+++ identityhub-service/src/main/java/br/dev/andrestamatto/identityhub/clientapplication/domain/MachineSettings.java
@@ -1,0 +1,9 @@
+package br.dev.andrestamatto.identityhub.clientapplication.domain;
+
+public record MachineSettings() implements ApplicationClientSettings {
+
+    @Override
+    public ApplicationClientType type() {
+        return ApplicationClientType.MACHINE;
+    }
+}